Initial tweaks to get it working
