+++ fastly/tls_test.go
@@ -1,0 +1,205 @@
+package fastly
+
+import (
+	"crypto/rand"
+	"crypto/rsa"
+	"crypto/x509"
+	"crypto/x509/pkix"
+	"encoding/pem"
+	"fmt"
+	"math/big"
+	rnd "math/rand"
+	"strings"
+	"time"
+)
+
+func generateKey() (string, error) {
+	reader := rand.Reader
+	bitSize := 2048
+
+	key, err := rsa.GenerateKey(reader, bitSize)
+	if err != nil {
+		return "", err
+	}
+	var privateKey = &pem.Block{
+		Type:  "PRIVATE KEY",
+		Bytes: x509.MarshalPKCS1PrivateKey(key),
+	}
+
+	bytes := pem.EncodeToMemory(privateKey)
+
+	return string(bytes), nil
+}
+
+func generateKeyAndCert(SANs ...string) (string, string, error) {
+	privKey, key, err := buildPrivateKey()
+	if err != nil {
+		return "", "", err
+	}
+
+	cert, err := buildCertificate(privKey, SANs...)
+	if err != nil {
+		return "", "", err
+	}
+
+	return key, cert, nil
+}
+
+func generateKeyAndMultipleCerts(SANs ...string) (string, string, string, error) {
+	privKey, key, err := buildPrivateKey()
+	if err != nil {
+		return "", "", "", err
+	}
+
+	cert, err := buildCertificate(privKey, SANs...)
+	if err != nil {
+		return "", "", "", err
+	}
+
+	cert2, err := buildCertificate(privKey, SANs...)
+	if err != nil {
+		return "", "", "", err
+	}
+
+	return key, cert, cert2, nil
+}
+
+func generateKeyAndCertWithCA(domains ...string) (string, string, string, string, error) {
+	caCert, caPEM, caKey, caKeyPEM, err := buildCACertificate(domains...)
+	if err != nil {
+		return "", "", "", "", err
+	}
+
+	privateKey, key, err := buildPrivateKey()
+	if err != nil {
+		return "", "", "", "", err
+	}
+
+	cert, err := buildCertificateFromCA(caCert, privateKey, caKey, domains...)
+	if err != nil {
+		return "", "", "", "", err
+	}
+
+	return key, cert, caPEM, caKeyPEM, nil
+}
+
+func buildPrivateKey() (*rsa.PrivateKey, string, error) {
+	key, err := rsa.GenerateKey(rand.Reader, 2048)
+	if err != nil {
+		return nil, "", fmt.Errorf("failed to generate private key: %s", err)
+	}
+	privKeyBytes, err := x509.MarshalPKCS8PrivateKey(key)
+	if err != nil {
+		return nil, "", fmt.Errorf("unable to marshal private key: %v", err)
+	}
+	keyBlock := &pem.Block{
+		Type:  "PRIVATE KEY",
+		Bytes: privKeyBytes,
+	}
+	return key, strings.TrimSpace(string(pem.EncodeToMemory(keyBlock))), nil
+}
+
+func buildCertificate(privateKey *rsa.PrivateKey, domains ...string) (string, error) {
+	now := time.Now()
+	serialNumber := new(big.Int).SetInt64(rnd.Int63())
+	template := &x509.Certificate{
+		SerialNumber: serialNumber,
+		Subject: pkix.Name{
+			SerialNumber: fmt.Sprintf("%d", serialNumber),
+		},
+		NotBefore:             now,
+		NotAfter:              now.Add(24 * 90 * time.Hour),
+		KeyUsage:              x509.KeyUsageCertSign,
+		ExtKeyUsage:           []x509.ExtKeyUsage{x509.ExtKeyUsageServerAuth},
+		BasicConstraintsValid: true,
+		SignatureAlgorithm:    x509.SHA256WithRSA,
+		IsCA:                  true,
+		DNSNames:              domains,
+	}
+
+	if len(domains) != 0 {
+		template.Subject.CommonName = domains[0]
+	}
+
+	c, err := formatCertificate(template, template, &privateKey.PublicKey, privateKey)
+	if err != nil {
+		return "", err
+	}
+	return c, nil
+}
+
+func buildCertificateFromCA(ca *x509.Certificate, privateKey *rsa.PrivateKey, caKey *rsa.PrivateKey, domains ...string) (string, error) {
+	now := time.Now()
+	serialNumber := new(big.Int).SetInt64(rnd.Int63())
+	template := &x509.Certificate{
+		SerialNumber: serialNumber,
+		Subject: pkix.Name{
+			SerialNumber: fmt.Sprintf("%d", serialNumber),
+		},
+		NotBefore:             now,
+		NotAfter:              now.Add(24 * 90 * time.Hour),
+		KeyUsage:              x509.KeyUsageCertSign,
+		ExtKeyUsage:           []x509.ExtKeyUsage{x509.ExtKeyUsageServerAuth},
+		BasicConstraintsValid: true,
+		SignatureAlgorithm:    x509.SHA256WithRSA,
+		DNSNames:              domains,
+	}
+
+	if len(domains) != 0 {
+		template.Subject.CommonName = domains[0]
+	}
+
+	c, err := formatCertificate(template, ca, &privateKey.PublicKey, caKey)
+	if err != nil {
+		return "", err
+	}
+	return c, nil
+}
+
+func buildCACertificate(domains ...string) (*x509.Certificate, string, *rsa.PrivateKey, string, error) {
+	now := time.Now()
+	serialNumber := new(big.Int).SetInt64(rnd.Int63())
+	template := &x509.Certificate{
+		SerialNumber: serialNumber,
+		Subject: pkix.Name{
+			SerialNumber: fmt.Sprintf("%d", serialNumber),
+		},
+		NotBefore:             now,
+		NotAfter:              now.Add(24 * 90 * time.Hour),
+		KeyUsage:              x509.KeyUsageCertSign,
+		ExtKeyUsage:           []x509.ExtKeyUsage{x509.ExtKeyUsageServerAuth},
+		BasicConstraintsValid: true,
+		SignatureAlgorithm:    x509.SHA256WithRSA,
+		IsCA:                  true,
+		DNSNames:              domains,
+	}
+
+	privateKey, key, err := buildPrivateKey()
+	if err != nil {
+		return nil, "", nil, "", err
+	}
+
+	c, err := formatCertificate(template, template, &privateKey.PublicKey, privateKey)
+	if err != nil {
+		return nil, "", nil, "", err
+	}
+	return template, c, privateKey, key, nil
+}
+
+func formatCertificate(certificate *x509.Certificate, parent *x509.Certificate, publicKey *rsa.PublicKey, privateKey *rsa.PrivateKey) (string, error) {
+	derBytes, err := x509.CreateCertificate(
+		rand.Reader,
+		certificate,
+		parent,
+		publicKey,
+		privateKey,
+	)
+	if err != nil {
+		return "", err
+	}
+	certBlock := &pem.Block{
+		Type:  "CERTIFICATE",
+		Bytes: derBytes,
+	}
+	return strings.TrimSpace(string(pem.EncodeToMemory(certBlock))), nil
+}
